requirements.txt
asgiref==3.8.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg[binary]==3.2.9
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.2.0
gunicorn
whitenoise